feat(modularization): S-12 extract token-usage monitor to web/modules/token-usage.js
Move ~930 lines of token usage monitor frontend logic from the monolithic
app.js into a dedicated ES module: summary cards, timeline canvas (bars +
cumulative budget lines + legend), model distribution pie chart, MCP tool
stats grid, per-call detail table with sort/search, period/agent filters,
collect button and resize handler. Tooltip and event wiring moved from
module-top-level IIFEs into initTokenUsage(). app.js retains only the
import + initTokenUsage() call + registerPage('tokenUsage') reference.

Verified: node --check, tsc build, 2873 vitest tests green, /modules/token-usage.js
HTTP 200, authenticated Playwright smoke (page.route mocks) on port 3431:
summaryCount=2, canvasW=954, modelDistCount=1, toolStatsCount=2, detailCount=1,
budgetCardCount=2, 0 JS pageerrors.
